Create AUTH Middleware
this example, the /admin/dashboard route requires both authentication and authorization. The authenticateToken middleware verifies the user's JWT token, and the authorizeUser middleware checks if the user has the 'admin' role. If both conditions are met, the user can access the admin dashboard route. Otherwise, they receive a 403 Forbidden response.